pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Andrii Zabolotin <fenics7511@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.2.0"}
django-bootstrap-v5 = "^1.0.11"
python-slugify = "^8.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
